feat: App directory
