+++ app/head.tsx
@@ -1,21 +1,7 @@
-import { useEffect, useState } from 'react';
-import { useTheme } from 'next-themes';
-import Head from 'next/head';
-
-type Props = {
-  title?: string;
-  ogImageText?: string;
-};
-
-const Meta = ({ title = 'NextJs Boilerplate', ogImageText }: Props) => {
-  const [mounted, setMounted] = useState(false);
-  useEffect(() => setMounted(true), []);
-
-  const { theme } = useTheme();
-
+const Head = () => {
   return (
-    <Head>
-      <title>{title}</title>
+    <>
+      <title>NextJs Boilerplate</title>
       <meta name="viewport" content="initial-scale=1, width=device-width" />
       <meta
         name="apple-mobile-web-app-status-bar-style"
@@ -31,14 +17,10 @@ const Meta = ({ title = 'NextJs Boilerplate', ogImageText }: Props) => {
         property="og:url"
         content="https://next-js-boilerplate-sable.vercel.app"></meta>
       <meta property="og:type" content="website" />
-      <meta property="og:title" content={title} />
+      <meta property="og:title" content="NextJs Boilerplate" />
       <meta
         property="og:description"
         content="nextjs boilerplate website with presets"
-      />
-      <meta
-        property="og:image"
-        content={`/api/og${ogImageText ? '?title=' + ogImageText : ''}`}
       />
       <meta name="twitter:card" content="summary_large_image" />
       <meta
@@ -49,23 +31,13 @@ const Meta = ({ title = 'NextJs Boilerplate', ogImageText }: Props) => {
         property="twitter:url"
         content="https://next-js-boilerplate-sable.vercel.app"
       />
-      <meta name="twitter:title" content={title} />
+      <meta name="twitter:title" content="NextJs Boilerplate" />
       <meta
         name="twitter:description"
         content="nextjs boilerplate website with presets"
       />
-      <meta
-        name="twitter:image"
-        content={`/api/og${ogImageText ? '?title=' + ogImageText : ''}`}
-      />
-      {mounted && (
-        <meta
-          name="theme-color"
-          content={theme === 'light' ? '#f9fafb' : '#222222'}
-        />
-      )}
-    </Head>
+    </>
   );
 };
 
-export default Meta;
+export default Head;
