+++ components/navbar.tsx
@@ -1,3 +1,4 @@
+'use client';
 import Link from 'next/link';
 
 import { useTheme } from 'next-themes';
