+++ styles/global.css
@@ -1,16 +1,3 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
-
-#nprogress .bar {
-  background: #bb86fc !important;
-}
-
-#nprogress .peg {
-  box-shadow: 0 0 10px #bb86fc, 0 0 5px #bb86fc !important;
-}
-
-#nprogress .spinner-icon {
-  border-top-color: #bb86fc !important;
-  border-left-color: #bb86fc !important;
-}
